Add manual updates for pfsync commands.
While I'm here, I should also note that, in the last pfsync commit,
I also changed the behavior of 'show run' and 'show start' so that they
only display the configuration when you are in the privileged (enabled)
state.  The reason for this is simple- now that 'enable' actually does
something useful, we also have a crypted password hash to hide.