Move snapshot date to 2007-01-17
This project has been lagging for a bit of of 2005 and all of 2006

Users need various fixes to run with any recent version of OpenBSD, and
unfortunately no release was made in 2006 to address these issues at all.

Here it is.  More additions are needed to bring up to date with current
OpenBSD functionality, and hopefully I can get to that in the next month.
This better than what was previous released as a snapshot.   Also I'd like
to open anoncvs so people don't have to wait for snapshots, and maybe it
will encourage others to help and pitch in with development more.  I could
really use ideas and help to bring this program further along.  I have not used
nsh in a professional capacity in the past two years, at least not to as
much of a degree as I used to.  This could change at any time, but lately
i'm just too busy on other things.

I still would like to continue with nsh but, obviously, without some
outside help, further enhancements will be rather slow.  Also, I get
a lot of requests from people who do not have very much sophistication
with the C compiler and the OpenBSD environment.  I would encourage people
in this situation to think about what they want to accomplish and try
anyways, because this is a very easy project to work on, and you might
enjoy it...  If you are happy with C and OpenBSD, then there are plenty
of things in here that need to be cleaned up.  The CLI parser can be
largely rewritten, ipsecctl, bgp, etc can all be added, either as
functionality where the ctl executable is called, or with direct kernel
calls.  In most cases, calling the external executable is going to be
preferrable.  If you are interested then talk to me and I can direct you..