make ctl.c routines more generic, and replace separate routines for each action with tables that contain the command names and argument lists.
uses genget in cmdhandler to sort with prefix/ambiguous support.  reuses code
from commands.c to parse the argument list from the table and fill space
allocated from the generic handler (pr_prot1, cmdhandler). this creates
step_optreq() which is currently dependant on the interpreted (typed or cmdrc)
command having the first two arguments never get copied over to the executed
argument list.  if you read the structures, then the drawbacks become
pretty obvious.

this is sort of a bug, or a cheap design trick but it fits the current usage
perfectly, so this note is just to remind me to think about it more in the
future as the usage of step_optreq() may become more sophisticated.