terminate all argument lists parsed through step_optreq() with NULL and not '\0'.  step_optreq() still terminates them with '\0' for execv() (as per the exec(3) man page) once it sees the NULL, or when it runs out of arguments to fill in.
in step_optreq, change the meaning of the last argument to 'skip' - that
is, how many arguments in the beginning of argv do we need to skip to get
to the optional arguments that we will fill in OPT or REQ arguments with.
now the number of non-OPT/REQ arguments in the compiled in argument
can be different than the number of arguments typed in (since they have
nothing to do with each other anyways, this flexibility makes sense)

this resolves the distaste with this function that I expressed in the last
commit message.  (step_optreq could be further improved to allow static
arguments to be mixed into optional argument lists, but given the limited
data available from the table, it's unclear to me how this would work in all
cases)