Dump StringList based makeargv and revert back to 4.4BSD makeargv, but with cursor_argc/cursor_argo tracking added (for completion)
The newer makeargv was painful to read, and didn't handle fgets supplied
lines properly.  It did have the advantage of dynamically allocating
space for new arguments through the use of sl_add.  While this was nice,
it really isn't necessary.  Arguments are always parsed from a char line[].
So, we just set the max arguments to sizeof(line)/2 which is impossible
to exceed (an argument must be at least one character, and then it must
have another charater, space or \t to delineate.)  This ties makeargv()
to line[], which, surprise, is already the case.