Fix trunkproto, now it always takes the right name and the config file always displays the right name
This is overly complicated. Why are the enumerators numbered differently than the order of the name elements in the trunk proto structrue itself? The bridge code (balancing names) handles this in the appropriate way, trunk does not.

As a side effect, trunk is actually usable via nsh now. This was broken for 6 years. Sorry.