show.c says "The address is assumed to be that of a net or subnet, not a host." for netname functions.
So, we use netname4 to create a CIDR notation for a host or network IP address. Using netname6 for
either does not work quite as nicely. It actively overwrites the host portions of the IPv6 address.
There is no good reason to do this, even when used in the original convention of netname6. If a
non-network address is used in the routing table, shouldn't we be able to obviously discern that flaw,
instead of papering over it?

This fixes the IPv6 address display problems for functions that already use netname6 (And shouldn't
break anything that actually wants the original netname function)