start using the local system installboot, not the destination system's copy. this makes it possible to build a system version on one end of the time_t jump while your host system is on the other end.
before we assumed that there was always going to be new magic in the
installboot which might be useful. now we assume there is new evil.
perhaps the shift back to assuming magic will resume in the future.
for now make life easier with the current situation as it is.