1. Get rid of all vestiges of C/H/S user interface and replace with -l <blocks> (sectors) for both flashrd and growimg. CHS values are placeholders, ignored for most purposes
2. Use new fdisk -l <blocks> option whenever the target is an image.
No CHS information will be saved on the image MBR when -l is used ("forces"
LBA mode). In a future release I may switch all fdisk invocation to use -l.
REQUIRES March 3rd or later 5.5 snapshot, OR rebuild fdisk from
recent sources prior to using flashrd/growimg!

3. Set cpg to 1 on disklabels (the kernel has been doing this for a while
anyways, might as well catch up to reality)

4. Block size can be specified to flashrd, growimg with -l blocks -b bytes
but this is not documented and NOT expected to work (just for diddling
at this time, it's not at all clear that fdisk or the OpenBSD kernel is
fully ready for block sizes != 512 bytes although perhaps disklabel and
newfs could be designed for 4K blocks on some devices while fdisk still
uses a sector count representing 512 byte blocks)

5. Stop trying to re-use ANY old disklabel information when converting
images with growimg. (Not sure how I ever thought keeping this was a worthy
idea.)