Move to uniform disklabel handling with disktab
1. Gives us the ability to specify disk geometry to vnconfig, which is
the only way to use vnd with sector sizes other than 512bytes

2. Disklabel handling is now the same throughout all scripts which
need to create a label. vnconfig AND disklabel -w always
get all exact information in the exact same way. No more trying to
save old disklable information such as duids, what a stupid idea
that was.

The only disadvantage is that we need write access to the host
system /etc to be able to move the system disktab out of place
and put our temporary one in place. Although it's not ideal, it
is the only way to use disktab with the current libc implementation
short of chroot or other tricks that are not worth the effort.