commit the newer /usr/src/usr.sbin/arp/arp.c in place of the 7yr old version!
something that changed made arp set not work at all, and updating the
whole thing (including the arp version of rtmsg() ) is the quickest
fix for now. there are some features here that we don't take
advantage of yet. this and other routing socket usage needs a careful
examination to factor out common code.