Always print 'no shutdown' for pppoe interfaces (until the kernel can be fixed.)
For some reason, the kernel messes with IFF_UP on pppoe during
protocol events, rather than IFF_RUNNING. It isn't immediately
obvious where the problem is with my cursory examination of
if_spppsubr.c and if_pppoe.c. Once this is fixed in the kernel,
this hack will be removed.