bpe and nvgre break the nsh mold, they are both interfaces and bridges, that is, they answer to SIOCBRDGRTS but also support ioctls that were, before around 2015, only used on interfaces
the whole point of separating interfaces and bridges was the huge command list for interfaces (and easing people into the the conceptually different types of interface)...this needs to be thought out in a different way so we only list relevant commands for each type of interface, since there are so many unique types now!!

but for the moment, we keep the current abstraction alive. to do this, we need to start supporting some interface style ioctls on bridges.

so, we start with 'vnetid' and 'parent' being added to bridge mode