directly set the IFXF_AUTOCONF4 flag if dhcpleased(8) is running
As of OpenBSD 7.2 dhclient is a stub which sets the IFXF_AUTOCONF4 flag
for dhcpleased. Prepare nsh for a future after rm /sbin/dhclient by
checking if dhcpleased is running and directly setting the flag if it is.

This is a first future-proofing step. Further changes will be required
for full dhcpleased support in nsh. For example, write-config will
currently write out a leased IP as if it was a static IP.

As a side effect this adds a new "autoconf4" command, which only works
on systems running dhcpleased. For now, I have left this undocumented
in the man page, since users can simply keep configuring DHCP as usual.
In the future, the man page should probably at least mention it, though,
because this command now appears in the user's tab-completion list.