Change 'config' to 'configure' (still accepts partial words, like all commands) and enforce it for commands in the global scope that make system configuration changes. It also accepts 'configure terminal' like Cisco IOS.
This is a simple guard against commands being used which change system
configuration until 'config' mode is entered. Since NSH targets users
who may not be extremely familiar or comfortable with a CLI environment,
this is useful. Accidental cut-and-paste is a thing! The same users may
likely to be familiar with 'config t' as cisco shorthand and it's
nice if we can act in a familiar manner.

This means that a majority of NSH commands now require both 'configure'
and 'enable' before they can be used. In the future, perhaps we'll find
that 'enable' is superfluous.