make the "autoconf4" command work on pppoe interfaces
dcpdleased ignores the AUTOCONF4 interface flag on point-to-point
interfaces such as pppoe. So setting this flag on pppoe interfaces
as nsh would do it has no effect.

As of this change the autoconf4 command instead configures wildcard
addresses and routes for pppoe interfaces. These wildcards will be
replaced with usable addresses by sppp(4) during IPCP negotiation.

IPv6 is not handled yet, this will be done later.