add tab-completion to "inet" and "inet6" commands and revive "ip"
Add a new tab-completion mechanism which displays usage information
when some or all parameters cannot be expanded automatically.
Parameters which are not arbitrary will be completed as usual.
For this to work, placeholders representing arbitrary parameters, such
as IP addresses, must be listed in pointy brackets in the help text.
Such arbitrary parameters will never be tab-completed.
Non-arbitrary parameters must be listed without brackets and will be
tab-completed if matched.
This enables full tab-completion support for commands like "inet dhcp"
which was impossible to implement with existing completion mechanisms.

Add the following new commands, all of which can be tab-completed:

  inet autoconf  (equivalent to autoconf4 command)
  inet dhcp      (likewise equivalent to autoconf4 command)
  inet6 autoconf (equivalent to autoconf6 command)

Also restore the "ip" command as an alias for "inet" such that users
who are often typing on Ciscos don't have to adjust their keystrokes.
This makes the following commands official again and also work with
tab-completion:

  ip autoconf
  ip dhcp