switch CMPL(h) to use a table of struct ghs instead of a table of strings
In the future this will allow completion of parameters which follow upon
parameters using CMPL(h) to complete themselves.

It also allows us to remove a hack which checked the struct size before
using ghs->table since this was previously an invalid pointer in case
of CMPL(h).

The previous code only stored a string for CMPL(h) completions,
abusing the fact that ghs->name is the first element of struct ghs.
Having struct ghs * pointers running around in the program which actually
point to memory objects smaller than struct ghs is a recipe for disaster.