make cmdargs() return the exit code of the called program
This makes it possible to detect whether a command we ran did signal
an error condition via its exit code. It will come in handy when
running commands to check syntax of config files, for example.

Currently no caller is checking the return code of cmdargs() so this
change is a no-op as of now.

If running the program in the child process fails ensure that a non-zero
exit code gets returned. I chose exit code 127 which matches the exit
code returned by ksh when a command cannot be found.