ensure that /etc/nshrc is not world-readable
This file may contain configuration settings that are only exposed to
the root user by the kernel. Saving the configuration to a file which
is world-readable could expose such information.

An example of such information are public wireguard keys, which are
not meant to be publicly distributed at large:

https://marc.info/?l=openbsd-tech&m=159384133419655&w=2

List:       openbsd-tech
Subject:    Re: get public key as non-root
From:       "Jason A. Donenfeld" <Jason () zx2c4 ! com>
Date:       2020-07-04 5:40:35
Message-ID: CAHmME9pf0a7zn9-eOmnOg94C+LcHh5Mqf_RqPJTo7JL-w4g8HA () mail ! gmail ! com

On Fri, Jul 3, 2020 at 11:47 AM Klemens Nanni <kn@openbsd.org> wrote:
> Is there any particular reason why an interface's *public* key is only
> shown to the root user in ifconfig?

Yes, there is a reason for this.

The WireGuard protocol has a property called "identity hiding". See
section 3.4 and 4.3.4 lemma 7 of
<https://www.wireguard.com/papers/wireguard-formal-verification.pdf>
or section 7.8 of
<http://www.noiseprotocol.org/noise.html#identity-hiding>. The mac1
value also relies on this identity hiding property. In other words,
public keys should not be easily broadcasted and should not be
accessible to unprivileged users.