use more(1) style input keys for the pager
Input keys accepted by the pager do not seem to be documented. Because most
people will be used to hitting keys used by more(1) and similar tools we
should use the same keys as offered by more(1) to advance to the next line
or page.

Enter, j, and Ctrl-n now advance to the next line.
Space, f, and Ctrl-f now advance to the next page.

We no longer make unrecognized keys advance to the next page because this
makes it harder to discover the set of valid keys by trial-and-error.