add a runat command and scripting via the nsh command-line
Add a runat command, which based on at(1), to allow nsh commands
to be run at some time in the future. The name "runat" instead of
"at" was chosen to avoid conflicts with future commands that begin
with "at" (like we saw when "pin" and "ping" when we had to rename
the former to "setpin" to disambiguate the two).

To allow runat to work without having to store nsh commands in a
temporary file which would never get removed, we now allow commands
to be piped into nsh's standard input, a.k.a non-interactive mode.
This way, nsh can be set as the SHELL for at(1) to run, and commands
will be fed into nsh's standard input by the at(1) job script.

There are commands which require interactive input and will not
work as intended in non-interactive mode.
Such commands will need to be made aware of non-interactive mode.
For now, make 'reboot' aware, such that planned reboots can be
scheduled without failure.

For ease of use, the runat command supports tab-completion for some
variants of the timespec syntax of at(1). It does not attempt to
tab-complete the list of commands, however, as this would be much
too complicated. Tab-completion also breaks as soon as the user
deviates from the set of suggested completions, which is an inherent
limitation of the current tab-completion system.

Some basic verification of commands is performed before a job is scheduled.
For now, we make sure that an internal or external command to run actually
exists, and that commands which require privileged mode are scheduled in
privileged mode.
Block the 'enable' command due to an issue where doas(1) and su(1) would
end up reading input that is meant to be read by the privileged nsh child
process. Hopefully we will find a solution to this at some point.