allow nsh to change its behaviour when standard input is not a TTY
Introduce an 'interactive_mode' flag which indicates whether standard
input is a TTY. This allows nsh to accept commands on standard input
for scripting purposes.

This will be needed for better integration of at(1).

The 'enable' command will not work in non-interactive mode yet.
This will be fixed soon.

ok tom