! without arguments should not fork a new shell
This Diff changes the behaviour of nsh interactively by checking if ! has been entered without any arguments, 
if the user is in global config mode or privileged mode the user is returned to the same prompt. if the user is in interface config mode or bridge config mode 
the user is returned to the global configuration  mode  This is to facilitate users copying and pasting nshrc configuration from one device to another interactively.